pyproject.toml
[project]
name = "django-cadastro"
version = "0.1.0"
description = ""
authors = [
    {name = "willian"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-gis (>=1.2.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
